+++ src/engine/LightSource.cpp
@@ -16,8 +16,8 @@ LightSource::LightSource(const Vector3& position,
 LightSource::LightSource(const Vector3& position,
                          const Vector3& uDir,
                          const Vector3& vDir,
-                         const double width,
-                         const double height,
+                         const float width,
+                         const float height,
                          const Vector3& colorDiffuse,
                          const Vector3& colorSpecular)
     : position_(position),
@@ -42,7 +42,7 @@ const Vector3& LightSource::getColorSpecular() const {
     return colorSpecular_;
 }
 
-double LightSource::getIntensity() const {
+float LightSource::getIntensity() const {
     return intensity_;
 }
 
@@ -52,15 +52,15 @@ GPU::GPULightSource LightSource::toGPU()
     data.position = glm::vec3(position_.x(), position_.y(), position_.z());
     data.colorDiffuse = glm::vec3(colorDiffuse_.x(), colorDiffuse_.y(), colorDiffuse_.z());
     data.colorSpecular = glm::vec3(colorSpecular_.x(), colorSpecular_.y(), colorSpecular_.z());
-    data.intensity = static_cast<float>(intensity_);
+    data.intensity = intensity_;
     data.uDir = glm::vec3(uDir_.x(), uDir_.y(), uDir_.z());
     data.vDir = glm::vec3(vDir_.x(), vDir_.y(), vDir_.z());
-    data.halfWidth = static_cast<float>(halfWidth_);
-    data.halfHeight = static_cast<float>(halfHeight_);
+    data.halfWidth = halfWidth_;
+    data.halfHeight = halfHeight_;
     return data;
 }
 
-void LightSource::setIntensity(const double& intensity) {
+void LightSource::setIntensity(const float& intensity) {
     intensity_ = intensity;
 }
 
@@ -71,8 +71,8 @@ void LightSource::setIntensity(const double& intensity) {
 //         return position_;
 //     }
 //
-//     const double rx = (static_cast<double>(rand()) / RAND_MAX) * 2.0 - 1.0;
-//     const double ry = (static_cast<double>(rand()) / RAND_MAX) * 2.0 - 1.0;
+//     const float rx = (static_cast<float>(rand()) / RAND_MAX) * 2.0 - 1.0;
+//     const float ry = (static_cast<float>(rand()) / RAND_MAX) * 2.0 - 1.0;
 //
 //     const Vector3 offset = uDir_ * (rx * halfWidth_) + vDir_ * (ry * halfHeight_);
 //     return position_ + offset;
